openzeppelin_monitor::services
Module blockchain

Blockchain client interfaces and implementations.

Provides abstractions and concrete implementations for interacting with different blockchain networks. Includes:

Generic blockchain client trait

EVM and Stellar specific clients

Network transport implementations

Error handling for blockchain operations

Client pool for managing multiple clients

Structs

ClientPool

Main client pool manager that handles multiple blockchain types.

EVMTransportClient

A client for interacting with EVM-compatible blockchain nodes

EndpointManager

Manages the rotation of blockchain RPC endpoints

EvmClient

Client implementation for Ethereum Virtual Machine (EVM) compatible blockchains

HttpTransportClient

Basic HTTP transport client for blockchain interactions

StellarClient

Client implementation for the Stellar blockchain

StellarTransportClient

A client for interacting with Stellar-compatible blockchain nodes

TransientErrorRetryStrategy

A default retry strategy that retries on requests based on the status code This can be used to customise the retry strategy

Enums

BlockChainError

Represents possible errors that can occur during blockchain operations

Traits

BlockChainClient

Defines the core interface for blockchain clients

BlockFilterFactory

Defines the factory interface for creating block filters

BlockchainTransport

Base trait for all blockchain transport clients

ClientPoolTrait

Trait for the client pool.

EvmClientTrait

Extended functionality specific to EVM-compatible blockchains

RotatingTransport

Extension trait for transports that support URL rotation

StellarClientTrait

Extended functionality specific to the Stellar blockchain