openzeppelin_monitor::utils::monitor::execution
Function execute_monitor

pub async fn execute_monitor<T: ClientPoolTrait, M: MonitorRepositoryTrait<N, TR>, N: NetworkRepositoryTrait, TR: TriggerRepositoryTrait>(
    monitor_path: &str,
    network_slug: Option<&String>,
    block_number: Option<&u64>,
    monitor_service: Arc<Mutex<MonitorService<M, N, TR>>>,
    network_service: Arc<Mutex<NetworkService<N>>>,
    filter_service: Arc<FilterService>,
    client_pool: T,
) -> ExecutionResult<String>

Executes a monitor against a specific block number on a blockchain network.

This function allows testing monitors by running them against historical blocks. It supports both EVM and Stellar networks, retrieving the block data and applying the monitor’s filters to check for matches.

Arguments

monitor_name - The name of the monitor to execute

network_slug - The network identifier to run the monitor against

block_number - The specific block number to analyze

active_monitors - List of currently active monitors

network_service - The network service to use

filter_service - The filter service to use

client_pool - The client pool to use

Returns

Result<String, ExecutionError> - JSON string containing matches or error